[파일분리] class별로 분리시켜보았다.
