--- raccoon-shopping-express/front/const.js
@@ -1,11 +1,0 @@
-const $mileageEventSlide = document.querySelector('.event--slide');
-const $prevEventButton = document.querySelector('.slide--button--prev');
-const $nextEventButton = document.querySelector('.slide--button--next');
-
-const $topMileageSlide = document.querySelector('#topMileageSlide');
-const $mileageSlidePage = document.querySelector('#mileageSlidePage');
-const $mallEventList = document.querySelector('#mallEventList');
-const $mallEventSlide = document.querySelector('#mallEventSlide');
-const $hotDealWrapper = document.querySelector('#hot-deal__wrapper');
-
-export { $mileageEventSlide, $prevEventButton, $nextEventButton, $topMileageSlide, $mileageSlidePage, $mallEventList, $mallEventSlide, $hotDealWrapper };
